services/web/requirements.txt
requests
python-dotenv
selenium
fastapi==0.55.1
uvicorn==0.11.5
fastapi-sqlalchemy 
pydantic 
alembic
psycopg2 
werkzeug
python-multipart
pyjwt